+++ lib/configuration/_all.dart
@@ -1,2 +1,3 @@
 export 'abstract_configuration.dart';
 export 'abstract_translations.dart';
+export 'refresh_configuration.dart';